+++ tasmota-lights-control-be/src/domain/profile.rs
@@ -4,10 +4,10 @@ use uuid::Uuid;
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
 pub struct LightInput {
     pub name: String,
-    pub dimmer: u8,
+    pub dimmer: i64,
     pub mode: String,
     pub rgb_color: RequiredOption<String>,
-    pub color_temperature_kelvin: RequiredOption<u16>,
+    pub color_temperature_kelvin: RequiredOption<i64>,
 }
 #[derive(Serialize, Clone)]
 #[serde(rename_all = "camelCase")]
